Fix archiver bug concerning ignored day directories.
The archiver sometimes incorrectly ignored day directories that were in
the "wrong" month. For example, if the day directory for April 30 was
placed in the May month directory (the norm for Old Bird data) and an
archiver end date of April 30 was specified, the entire May month
directory was ignored, even though the April 30 day directory should not
have been. The fix was never to ignore an entire month directory, but
rather to consider each day directory independently.