Modify viewer to advance across pages after classification.
After a singleton classification of the last clip on a page, the viewer
now advances to the next page (if there is one) and selects the first
clip on the new page. The viewer also advances to the next page if all
of the clips on the current page are classified with a single command,
but does not select the first clip of the new page.