Refine archive calendar color matching.
For some reason, on Mac OS X if one sets the face color of a Matplotlib
figure to the default PyQt4 window color the two resulting colors do not
match. They do match on Windows, however. So we now behave differently
on the two platforms so the colors match on both.