Added extension manager and improved VCL help.
The extension manager provides a way for any part of an application to
get the extensions that are present for a given extension point.

Improved VCL help so that commands learn about their subcommands from
the extension manager rather than knowing about them a priori. Added a
"help" command. Added infrastructure to support arguments common to many
commands and to format help. Expanded help text for all commands.