Changed how viewer --archive command line argument works.
The argument value now specifies an archive directory path rather than
an archive name, and defaults to the current working directory. The
"archives" and "archive" viewer preferences are no longer supported.
This makes the viewer --archive command line argument just like that of
VCL.

Also improved the error message a user sees if they specify an archive
directory path that does not exist.