Added --preferences option to viewer.
In order to make this work, it was necessary to change the way
preferences are accessed. They are no longer stored in a public
dictionary in the `prefs` attribute of the preferences module, but
instead are stored in a private dictionary. A preference value is now
obtained by calling the `get` function of the preferences module rather
than by simply accessing a value in the preferences dictionary.