Changed viewer presets format YAML.
Previously is was assumed that presets could be stored in various
formats, depending on the preset type, including JSON and custom
formats. We now assume that all presets are stored in YAML format.

Also modified preset manager and preset manager unit tests so that the
tests do not produce text output. The appearance of such text output
when running the unit tests was confusing.