Made lat/lon order consistent throughout the project.
The order was formerly inconsistent, though mostly I put longitude
before latitude. After some reading on the web I gather that both of the
possible conventions are used. The traditional navigational convention
and the convention that is usually recommended for data display is
latitude before longitude. However, on maps longitude is usually
horizontal and latitude vertical, and the convention with cartesian
coordinates is to put the horizontal coordinate before the vertical one,
which would suggest that longitude come before latitude. I have chosen
to put latitude before longitude in displays to respect the usual
convention, and to the extent possible to use the same convention
elsewhere (for example, for function arguments and data storage) for
consistency.