Corrected an integer division error for Python 3.
A division operation in the time/frequency analysis utilities involving
two integers should have been a floor division (using the `//` operator)
but was a true division (using the `/` operator) instead. In this
particular case the result of true division has the correct numerical
value in both Python 2 and Python 3, but its type is `int` for Python 2
and `float` for Python 3. The type needs to be `int` since the result is
used as a NumPy array dimension, so I have changed the division to floor
division.