Moved mpg_ranch and old_bird packages into vesper package.
This puts all Vesper code into a single package that can be distributed
via PyPI or Anaconda.org. I had made separate top-level packages for MPG
Ranch and Old Bird code since I had in mind distributing those as
separate plugins. We may still do that, in which case the code may move
out of the `vesper` package again, but we can cross that bridge when we
come to it.