Revised audio recorder class to use a ring of input buffers.
Audio data delivered by PyAudio via the callback function are copied
into the buffers before being sent to the recorder thread. The audio
data buffers delivered to the callback function used to be sent directly
to the recorder thread, a bad idea since we don't know what PyAudio does
with a buffer after the callback returns.