Finished implementing "no annotation" choice in clip collection view.
For now this choice is presented with the name "Unclassified", but this
will need to be generalized when we support annotations other than
classifications.

The basic query for the night view is now of the clip table rather than
the annotation table, and we retrieve all annotations for each clip
separately. We may want to optimize this in the future to avoid so many
queries, but I figured I'd start with the obvious implementation.