Reduced night view load time.
The changes in this commit reduced the load time for a 6700-clip night
from about 20 seconds to about 2 seconds.

There are two changes, each of which reduced the number of database
queries needed for each clip at load time. The first change was to defer
the query that gets a clip's annotations until the clip is actually
displayed. The second thing was to temporarily hard code the clip sample
rate. We will fix the hard-coded sample rate in a future commit.