Added `INCLUDE_DJANGO_DEBUG_TOOLBAR` Vesper Django app setting.
Whether or not the Django Debug Toolbar was included used to be
controlled by the `DEBUG` setting. That setting must be `True` for
deployments on the Django test server (yes, we know that's not
recommended: we will begin deploying to nginx/uWSGI soon!) in order to
serve static content correctly, so we've split out control of the Django
Debug Toolbar so that it can be excluded from deployments on the Django
test server.