Fixed multi-clip classification bug from recent Vesper Workshop.
We discovered at the workshop that an attempt to classify all of the
clips of a clip album page with one keystroke resulted in "database is
locked" errors. Unfortunately the only solution with SQLite appears to
be to use our own lock to ensure that we attempt no more than one
database transaction at once. As a quick fix, Vesper currently uses the
lock only for manual classification. It will use it for all database
writes in the near future.