Modified clip album to ignore redundant classification commands.
If the user issues a command via the keyboard to reannotate a clip with
an annotation value that it already has, or to delete an annotation that
it does not have, the command is now ignored. Formerly it was not. The
change eliminates some unnecessary HTTP requests, and simplifies
annotation edit histories.