Modified clip album rug plot to always show entire night.
Previously, the rug plot limits were set according to the recording time
limits. After some use, however, it seems clear to me that it is
preferable to always show the entire night. The consistency makes the
plots easier to interpret, and it makes it easier to see what portion of
a night was covered if recording was interrupted.