Improved recorder input overflow handling.
The recorder now logs two error messages for each (maximal) sequence of
consecutive input buffers for which input overflows. The first message
simply indicates that overflow is occurring, and is logged at the
beginning of the incident. The second message indicates the extent of
the incident, and is logged when the incident ends.

The changes fix a bug that caused the callback thread to block while
waiting for an input buffer to become available. The callback thread is
very time-sensitive and should never block, so now we simply discard
input samples (and report overflow) if the callback is invoked when no
input buffers are available.