Improved recorder channel number inference for recording imports.
The import recordings command used to assume that if recorder channel
numbers were not specified in the name of an imported file, the mapping
from recording channel numbers to recorder channel numbers was the
identity. This failed recently for monaural recordings from the MPG
Ranch that involved a single microphone attached to a recorder's input 1
(rather than input 0). The importer now assumes that if the number of
recording channels matches the number of recorder inputs to which
microphones are connected, the recorder inputs are associated with
recording channels in order of increasing input/channel number. This
closes issue #109.