Added support to clip album rug plot for missing solar events.
This commit addresses issue #94. At sufficiently high latitudes a night
may lack certain solar events. The rug plot used to assume that all
events were present, but it now handles missing events properly. (There
is one small caveat. The rug plot assumes that if all solar events are
missing for a day, the sun is up. This is correct near the summer
solstice, but not near the winter solstice. I do not expect this to be
an issue for nocturnal bird migration monitoring, the current focus of
Vesper. If it becomes an issue for some other type of monitoring, we can
deal with it.)