Added initial version of more general clip album.
The existing clip album displays all of the clips for a particular
station/mic, detector, classification, and night. One of the new clip
albums displays all of the clips for a particular station/mic, detector,
and classification, regardless of night. Future versions of the new clip
album may support a wider range of queries.