Switched to storing relative paths for recording files.
Formerly we stored absolute recording file paths in Vesper archives. We
now store only relative paths, and make those paths absolute as needed
with the help of a list of recording directory paths in the archive
settings. This will make it easier to relocate both recordings and
archives. A new command yet to be implemented will replace absolute
recording file paths in existing Vesper archives with relative ones.