Refactored client code some more, especially DFT.
Refactored the DFT module, converted unit tests to run under Jasmine,
and created some additional tests. One of the new tests had a bug that
caused it to hang. The resulting behavior was inconsistent since Jasmine
runs tests in parallel, so it took me a long time to find the bug.