Revised clip album keyboard input interpreter.
The keyboard input interpreter used to be called the command
interpreter. The revisions primarily make it easier to implement UI
components that respond to keyboard input. They introduce the notion of
a *commandable* object, and the notion of a single *commandable
delegate* object that can be shared by any number of commandable objects
and execute commands (bound to key sequences) on those objects.