Fixed bug concerning spectrogram clip view frequency axis.
The bug caused the spectrogram view and the time-frequency point overlay
to miscalculate frequencies from canvas y coordinates when the sample
rate was not 11025 hertz. The bug was introduced recently, when new
spectrogram settings were devised. It was discovered while preparing to
mark call centers in a collection of clips, before the marking actually
started.