Improved clip calendar and clip album Safari compatibility.
Unlike in other browsers, in Safari an ES2015 module in a JavaScript
file included in an HTML file via a <script> tag does not see global
variables defined in inline JavaScript in the same file. We used to
depend on this in the clip calendar, night, and clip album pages to pass
information through the HTML template to JavaScript. The inline
JavaScript is now an ES2015 module that imports an init function from
the module in the separate JavaScript file and calls the function with a
state object containing the required information.

Safari also doesn't support the Web Audio API as fully as other browsers
yet. This is partly, though not fully, addressed by this commit. Some
work remains to be done both to get clip audio data from the server to
the client without having to increase its sample rate to 44100 hertz or
more, and to play clips. See comments in clip-manager.js for more about
this.