Made detect command compatible with multi-threshold detectors.
A multi-threshold detector can perform detection for multiple detection
thresholds at once, improving efficiency in some situations, for example
for the computation of precision vs. recall curves. A multi-threshold
detector specifies the detection threshold at which a clip was detected
in each call to its detection listener: the listener of the detect
command did not used to accept the threshold in its `process_clip`
method, but it now does.