Completed 1.0 versions of PNF tseep and thrush energy detectors.
These detectors were referred to as the "PNF 2018 Basic" detectors in
the previous commit. Also included are PNF 2018 baseline detectors.
These are derived from the energy detector and have precision-recall
curves that are very similar to those of the original Old Bird
detectors, except with their post-processing (including clip merging and
suppression) removed.