Added partial support for on-the-fly clip extraction.
This feature extracts clip samples from recording files when those are
available but clip files are not. Clip extraction is performed by the
new clip manager singleton. We will eventually move all responsibility
for dealing with clip files from the clip model class to the clip
manager. This commit begins that process.