Added experimental command that runs detectors on random station-nights.
The new command selects a random set of station-night combinations for
which there are recordings and runs detectors on all recordings for
those nights. Detection can be restricted to sub-intervals of recordings
by a *detection schedule*. The command defers creation of clips in the
archive for a particular recording file interval until detection
completes on that interval, at which point it creates all clips for the
interval in a single database transaction. This is much faster than
creating each clip in a separate transaction. The command does not
create clip sound files: it creates only database clip records.