Removed some detector and classifier extensions from Vesper.
The removed detectors are the new non-learning detectors added in Vesper
0.4.0, which have been deemed to need further development.

The removed classifiers are the original MPG Ranch coarse and species
classifiers. These are no longer in use and I do not wish to try to
maintain them. There appear to be problems with them associated with old
scikit-learn pickle files that don't seem worth worrying about.

The code for the removed detectors and classifiers is still present:
only the extension listings in the `singletons` module were removed.

It may be advantageous to remove the code for the removed classifiers
before long, for example after we have created a new species classifier.
Aside from eliminating some unused, broken code, this would allow us to
eliminate scikit-learn from the list of Vesper dependencies.