Fixed bug in Old Bird detector runner.
When a clip produced by an Old Bird detector was sufficiently close
(within two seconds) of the start of a file, the detector monitor tried
to access file samples with negative indices in its search for the
samples in the file. It now limits the search to an interval that starts
at a nonnegative index.