Implemented batch clip loading.
A client can now request batches of clip audios and clip annotations
from the Vesper server instead of having to request them one at a time.
This increases the speed at which a user can page through a clip album,
and it drastically reduces the number of HTTP requests that a client
sends to a server for each new clip album page.