Modified clip manager to load pages sequentially.
When asked to load multiple pages, the clip manager used to send
requests for all of the clips and annotations of the pages to the server
in quick succession. This sometimes caused the first page to load more
slowly because responses for some of the data for other pages were
received before responses for the data for the first page. The clip
manager now sends requests for the data for each page separately, and
waits for responses before sending requests for the data for the next
page. The new code uses promises heavily, including the async and await
keywords.