Modified new coarse classifiers to extract all needed clip samples.
The classifiers used to first get all of the samples of a clip as
specified by the detector, and then extract the samples needed for
classification from those. Sometimes, however, the clip as specified by
the detector did not always contain all of the samples needed for
classification, which caused classification to fail. The classifiers now
get the samples needed for classification directly from the clip
manager.