Modified old MPG Ranch coarse classifier to use tensorflow.keras.
The classifier used to use the plain keras package, since it was
developed before keras was incorporated into TensorFlow. Using
tensorflow.keras will allow us to eliminate the keras package as a
dependency. The Conda keras package has tended to lag behind the Conda
tensorflow package, so that will be a welcome change.

Also moved an OpenMP issue workaround from the new coarse classifier
training script to its own utility module so we could call it from the
old and new training scripts and classifiers.