Improved clip album paging responsiveness.
The clip manager used to respond to each call to its `pageNum` setter by
immediately sending requests to the server for any clip batches needed
(but not already fetched) for the specified page and, if relevant, for
its neighbors. If the user advanced quickly over many pages, for example
by holding down the appropriate key, the client sent many requests to
the server for clip batches that were not actually displayed, and when
the user finally stopped on a page, it took awhile for the many requests
that had been sent to complete, including the ones for the final page.

The clip manager now sends requests to the server for only one page at a
time (and, if relevant, its neighbors), remembering the most recently
set page number but not initiating requests for it until any pending
requests have completed. This makes the clip album much more responsive
in the above scenario.