Added BirdVoxDetect code.
BirdVoxDetect (https://github.com/BirdVox/birdvoxdetect) is a product of
the BirdVox project (https://wp.nyu.edu/birdvox). At least for the time
being, we have added the BirdVoxDetect code to Vesper as the
vesper.birdvox.birdvoxdetect_0_1_a0.birdvoxdetect package so that Vesper
users will be able to install BirdVoxDetect and all of its dependencies
just by installing the Vesper Conda package. (Currently, BirdVoxDetect
is only offered as a pip package, while Vesper is only offered as a
Conda package.)