Removed minimum separation criterion from MPG Ranch detectors.
The way the criterion was implemented, it made the detectors not
monotonically non-decreasing with decreasing thresholds, in the sense
that the clips produced for a higher threshold were not necessarily a
subset of the clips produced for a lower threshold. It is possible to
implement the criterion while retaining monotonicity, but it is
relatively difficult and computationally expensive. I decided to just
jettison the criterion, at least for now.