Modified YAML utils to load and dump YAML text using ruamel_yaml.
This is part of an effort to switch from using the PyYAML module to
using the ruamel_yaml module in all of our Vesper code. PyYAML loads and
dumps YAML according to the YAML 1.1 specification, while ruamel_yaml
can load and dump according to either YAML 1.1 or YAML 1.2. PyYAML is
also no longer under active development, while ruamel_yaml is. We want
to use YAML 1.2 to parse schedules, since we want times like "12:34:56"
to be parsed as strings rather than as sexagesimal numbers.