Added a new BirdVoxDetect version.
The version is called "BirdVoxDetect 0.1.a0 AT 05a". It is identical to
the existing version with the same name but without the final "a". I
added it since I needed to re-run that detector on some recordings
without affecting the clips from the existing detector. I'll probably
want to delete the new version, and perhaps others as well, in the
not-too-distant future.