Added false positive threshold to coarse classifier evaluation mode.
The threshold limits clips reclassified by the classifier as false
positives to those whose scores are at least the threshold. This
threshold complements the false negative threshold that was already in
place.

We also now allow threshold values of `None` that turn off false
negative and/or false positive reclassification.