Improved transfer call classifications command detector name matching.
The command used to require an exact match of a detector name with a key
of the `_CALL_START_WINDOWS` mapping. It now requires only that a key be
a prefix of the detector name. This allows multiple versions of a
detector to be covered by a single mapping item.