Edited files to reflect new convention regarding compound nouns.
Compound nouns that combine words like "clip", "recording", "preference"
and "preset" with words like "file" and "directory" will keep the first
noun singular. For example, a file that holds preferences will be called
a "preference file" rather than a "preferences file" and a directory
that holds clips will be called a "clip directory" rather than a "clips
directory". We were formerly inconsistent about this. The decision to
keep the first noun singular was made because (1) that seems to be the
usual convention in English (e.g. "grocery list", "apple tree",
"toolbox", "bookshelf") and it avoids a double-plural awkwardness in the
plural forms of the compound nouns (e.g. "recording directories" instead
of "recordings directories").