Fixed bug in recording importer.
The bug caused the importer to silently ignore unimported files that
were contiguous with a preceding file or files that had already been
imported. For some reason this issue arose for the 2016 MPG Ranch
archive, but it would seem that it is probably rare.

Also improved the efficiency of the importer. It used to parse all of
the recording files of an archive: now it parses only those to be
imported.