Enhanced relative start time measurement.
The measurement was formerly relative to the recording start time. It
can now be relative to the recording start or end time or any of ten
solar events. The result can also be negated to get the time before a
reference time instead of a time after.

Also fixed bug in DurationFormat.