Renamed relative time measurement value formatter.
The old name was "Time Difference Formatter". The new name is "Relative
Time Formatter", which is more consistent with the names of the
measurements the formatter is used with (e.g. "Relative Start Time" and
"Relative End Time") and also other time formatters (e.g. "Local Time
Formatter" and "UTC Time Formatter").