Add support for BirdVoxDetect adaptive thresholds.
Vesper now uses the `run_birdvoxdetect` script of the `vesper_birdvox`
package to run BirdVoxDetect, instead of running BirdVoxDetect directly.
This was needed to support adaptive thresholds for both present and
older versions of BirdVoxDetect, and also made it possible to move
BirdVoxDetect checklist processing from the `vesper` package to the
`vesper_birdvox` package, of which it will be easier to create new
releases.