Revised singletons.
Vesper's singleton instances used to be instances of the
vesper.util.singleton.Singleton class and live in the vesper.singletons
module. The Singleton class is now gone, and each singleton instance
lives in its own module in a new vesper.singleton package.

As part of this effort, in order to avoid import cycles I modified the
extension manager to load extensions lazily by type, i.e. to load the
extensions of each type only if and when they are requested, instead of
all at once. This is also more efficient, in some cases avoiding loading
many extensions that are never used.