Reload on clip filter modal OK press when filter doesn't change.
We used to do nothing if the user pressed the "OK" button in the "Set
clip filter" modal and the specified filter was the same as the current
one. However, it seems desirable to at least reload the page to bring it
up to date with any clip metadata changes, e.g. classification or tag
changes.