Revise Django project and app startup.
These changes are part of an effort to Dockerize Vesper, and to adopt
Django best practices. With the changes, Vesper now relies more heavily
on the Django project settings file for configuration, and it uses
the third-party `environs` module to parameterize the configuration with
environment variables. These environment variables can be set in a
.env file, a Dockerfile, or a docker-compose.yml file.