Remove pip requirements.txt file.
Having this file in the repository causes GitHub to automatically
generate pull requests to increment package version numbers in
the file whenever security issues are fixed in Vesper dependencies.
I think it will be better to just generate a fresh requirements.txt file
whenever we build a new Vesper Docker image. Instructions for how to do
that are now in the Vesper Dockerfile.