Fix bug concerning string annotation UI values.
The bug caused the `Archive` methods that get visible string annotation UI values and value specs to return `None` instead of a tuple of default values or value specs (for all clips, unannotated clips, and annotated clips), which caused the clip calendar to raise an exception when the `Classification` annotation did not have an annotation constraint. Fixes issue #226.