Modify Django view that creates LRGV clips.
The view now considers Nighthawk clip metadata to be the authority on recording lengths, and updates the length and end time of a recording when the Nighthawk clip metadata indicate something other than what is already in the archive database. This sometimes happens when both Old Bird detector clips and Nighthawk clips are archived, since the metadata for Old Bird detector clips contains only an estimate of the recording length.