Change implementation of recorder S3 file uploader.
The uploader used to be of type AudioFileProcessor, which was a type of subprocessor of an audio file writer. The uploader is now a regular Processor, like an audio file writer, which allows us to eliminate the AudioFileProcessor type.

As part of this effort, added a `finished` argument to the `Processor.process` method and removed the `Processor.stop` method. The `finished` argument was needed to allow an audio file writer to close its final audio file in the `process` method instead of the `stop` method and output its path. After that change, the `stop` method was no longer needed.

This commit clears the way for S3 audio file upload retries, which will follow soon.