Provide context to recorder processors.
The context is a way of providing information to the processors of a graph that is not available prior to runtime, and hence cannot be provided via processor settings. Examples of such information include a multiprocess logging queue and a set of processor types. All of the processors of a graph share the same context.