Improve S3 file uploader robustness.
The uploader now runs as a sidecar process that runs as long as the recorder is running, regardless of whether or not recording is in progress. It used to run only when recording was in progress. The uploader also now saves its state for a particular upload directory to a file in that directory, including a list of files that have been uploaded and a list of files for which one or more upload attempts have failed. The uploader reads the file before each consideration of the contents of the directory and writes the file afterward. This allows the state to persist across recorder instances, and also makes it possible to leave uploaded files in place if desired without much impact on uploader performance. Previously, the uploader would attempt to upload an already-uploaded file if the file was left in place.