Revise version module.
The module now includes a `__version__` attribute that is the authority for the `vesper` package version. The module parses this attribute to get its parts. I made this change in preparation for switching from using setuptools to Hatchling for the build system, including a switch from using a `setup.py` file to using a `pyproject.toml` file.